Move autocast file to query package
